refactor(onboarding): rendre la page bienvenue côté serveur
